+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New features
 
+* `.GithubRelease()` added to create a GitHub (pre)release [#29]
 * `.Rprofile()` gains a `path` argument (with defaults) for identifying an `.Rprofile` location [#23]
 * `.FindRprofile()` added to find your `.Rprofile` file; used within `.Rprofile()` [#23] 
 * Adds `.LintFile()` for selecting individual files to `lint` (defaults to currently opened file) [#21]